Fixed most of the gcc 3.3 warnings

+++ src/celengine/location.cpp
@@ -151,7 +151,7 @@ static void initFeatureTypeTable()
 {
     featureTableInitialized = true;
 
-    for (int i = 0; i < sizeof(FeatureNames) / sizeof(FeatureNames[0]); i++)
+    for (int i = 0; i < (int)(sizeof(FeatureNames) / sizeof(FeatureNames[0])); i++)
     {
         FeatureNameToFlag[FeatureNames[i].name] = FeatureNames[i].flag;
     }
